Create Code Breakers
This code is a way for a person to encrypt a series of words in a sentence or multiple sentences.  You have the option to enter your own input for the code to encrypt (I have the input commented out) so anyone can put whatever they want as the two key phrases.