Create Midterm Game
For a midterm in my introduction to computers class we were put into groups and assigned to create anything we wanted.  My group wanted to create a butterfly effect game, this is my portion of the game that I created.  Essentially a random location was chosen (my location was an island) then from my portion of the game I chose to use the random function to select random functions to take place.